Cleanup, include debug helpers and tests
