+++ treeflow/debug/minimize_eager.py
@@ -123,27 +123,3 @@ def fit_surrogate_posterior(
         optimizer=optimizer,
         **opt_kwargs
     )
-
-
-def _any_nonfinite(x):
-    nonfinite = tf.logical_not(tf.math.is_finite(x))
-    return tf.reduce_any(nonfinite)
-
-
-class NonfiniteConvergenceCriterion(
-    tfp.optimizer.convergence_criteria.ConvergenceCriterion
-):
-    def __init__(self, name="NonfiniteConvergenceCriterion"):
-        super().__init__(min_num_steps=0, name=name)
-
-    def _bootstrap(self, loss, grads, parameters):
-        return None
-
-    def _one_step(self, step, loss, grads, parameters, auxiliary_state):
-        loss_nonfinite = _any_nonfinite(loss)
-        grads_nonfinite = [_any_nonfinite(x) for x in grads]
-        parameters_nonfinite = [_any_nonfinite(x) for x in parameters]
-        has_converged = tf.reduce_any(
-            [loss_nonfinite] + grads_nonfinite + parameters_nonfinite
-        )
-        return has_converged, None
